Change from long to size_t to match with bulk

+++ include/options.hpp
@@ -2,21 +2,21 @@
 
 namespace pmondriaan {
 
-enum class m : long { cut_net, lambda_minus_one };
-enum class bisection : long { random, multilevel };
-enum class sampling : long { random, label_propagation };
+enum class m : int { cut_net, lambda_minus_one };
+enum class bisection : int { random, multilevel };
+enum class sampling : int { random, label_propagation };
 /**
  *
  */
 class options {
   public:
-    long sample_size;
-    long coarsening_max_clustersize;
-    long lp_max_iterations;
-    long coarsening_nrvertices;
-    long coarsening_maxrounds;
-    long KLFM_max_passes;
-    long KLFM_par_number_send_moves;
+    size_t sample_size;
+    size_t coarsening_max_clustersize;
+    size_t lp_max_iterations;
+    size_t coarsening_nrvertices;
+    size_t coarsening_maxrounds;
+    size_t KLFM_max_passes;
+    size_t KLFM_par_number_send_moves;
 
     m metric;
     bisection bisection_mode;
